add "file editor" plugin
This plugin allows editing template.html and slideviewer/custom.css
(or other files) directly from within the admin frontend.

Note that no security checks whatsoever are performed - you can edit
any text file on the system the webserver has access to by fiddling
around with the URL. This is still an admin-only feature, and we
assume that admins know what they are doing.