implement PartyMeister-style automatic live voting
The slideviewer can now emit "callbacks" after a specified (and configurable)
amount of time spent on a slide. This can in turn be used by the livevoting
plugin to enable livevoting for a specific entry after its beamslide has been
up for a few seconds. It's also possible to enable normal voting after the
"end of compo" slide has been shown. All of this is optional, of course.