allow changing the order in which compos are listed
This adds a new setting 'compo_order' that specifies what to put into the
ORDER BY clause then querying the database for a compo list. The default
is "start,id", which is the same value as always, but due to it being a
setting, it can be changed if needed. (There is no special UI for it though.)

The use case is exerting more control over the order of the compos as presented
in the admin UI, results etc. The "sort by start time" approach is already a
good start, but it falls a bit short in the presence of compo blocks where a
bunch of compos start at the same time.
The new compo_order setting can be used in two ways to enforce a custom order:
- Create all the compos upfront in the desired order and set compo_order to "id".
- Do "ALTER TABLE compos ADD order INT", set compo_order to that new column, and
  fill it with whatever numbers you like in Adminer.