smarter compo/prizegiving title slide display
The slideviewer currently displays "The XXX compo" in the
compo and prizegiving title slides. This commit inhibits
this behavior if the compo name itself already ends with
"compo" or "demoshow"; in these cases, the bare compo
name without pre- and postfixes will be used.

This behavior can also be enforced for *all* compos by
adding the global setting "slideviewer_barenames" with a
non-empty value.