new plugin: Warn-O-Matic - warn user if entry description text is too long
Compo slides only have limited space for the description text, and since
we don't have an auto-text-size mechanism like some other party systems,
excess text will break the layout or just be truncated at some point.
Since we also don't have slide previews, the user has no way of knowing
that this is going to happen.

To mitigate the issue, this plugin adds a bit of JavaScript to the entry
upload and update pages on the visitor website that adds a warning and
changes the color of the text field when too many lines of text are entered.
The number of lines is estimated by simply assuming that every line fits
a certain number of characters.
There are two warning levels: "warning" (yellow) for when the text starts
to fill up most of the available space, but is still expected to fit, and
"maximum" (red) when it's pretty clear that things are going to break.
The three parameters (assumed number of characters per line, and the
warning and maximum thresholds) can be adjusted in the settings editor
when the plugin is activated.